@@ -1,7 +1,7 @@
 {
   "name": "statuses",
   "description": "HTTP status utility",
-  "version": "2.0.1",
+  "version": "2.0.2",
   "contributors": [
     "Douglas Christopher Wilson <doug@somethingdoug.com>",
     "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
@@ -20,17 +20,17 @@
     "LICENSE"
   ],
   "devDependencies": {
-    "csv-parse": "4.14.2",
-    "eslint": "7.17.0",
+    "csv-parse": "4.16.3",
+    "eslint": "7.19.0",
     "eslint-config-standard": "14.1.1",
-    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-import": "2.31.0",
     "eslint-plugin-markdown": "1.0.2",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-promise": "4.3.1",
     "eslint-plugin-standard": "4.1.0",
-    "mocha": "8.2.1",
+    "mocha": "8.4.0",
     "nyc": "15.1.0",
-    "raw-body": "2.4.1",
+    "raw-body": "2.5.2",
     "stream-to-array": "2.3.0"
   },
   "engines": {
